+++ constants/strings.py
@@ -1,0 +1,29 @@
+NEW_USER_MESSAGE = "Hello! This is the first time we have seen your number. We'll need to do some setup. First, what is your name?"
+
+SETUP_MESSAGE2 = "Are you a buyer or a seller? Respond with 1 for buyer, 2 for seller."
+
+SETUP_ERRORMESSAGE2 = "Please respond with 1 or 2."
+
+SETUP_MESSAGE3 = "What county are you located in?"
+
+SETUP_MESSAGE4 = "Thank you! You're all set up. If you want to buy produce, text #BUY, if you'd like to sell produce text #SELL."
+
+REPORT_MESSAGE1 = "What crop are you selling?"
+
+REPORT_MESSAGE2 = "How many kilograms of your crop are you selling?"
+
+REPORT_MESSAGE3 = "Where are you located?"
+
+REPORT_MESSAGE4 = "What is the price of your crops?"
+
+REPORT_MESSAGE5 = "Thanks! Your crops have been posted. We will inform you when someone has purchased your crops."
+
+LIST_MESSAGE1 = "What crop are you looking for?"
+
+LIST_MESSAGE2 = "Where are you located?"
+
+LIST_MESSAGE3 = "What quantity of your crop do you want? Please list in kilograms."
+
+LIST_MESSAGE4 = "What is the maximum price you are willing to pay per kilogram in shillings?"
+
+LIST_MESSAGE5 = "These are the following posts that match your preferences. Please text the number associated with that post if you are interested in purchasing it."
