+++ objects/usercache.py
@@ -1,0 +1,6 @@
+class usercache():
+    def __init__(self):
+        self.eventType = 0
+        self.location = [0.0, 0.0]
+        self.events = []
+
